refactor: post feature ui 폴더 구조 정리
