+++ src/app/posts/create/page.tsx
@@ -2,7 +2,7 @@
 
 import { POST_TAB_VALUES, POST_TABS } from '@/features/posts/constants/post-tabs';
 import { useCreatePost } from '@/features/posts/hooks/use-create-post';
-import PostForm from '@/features/posts/ui/post-form';
+import PostForm from '@/features/posts/ui/post-form/post-form';
 import type { PostFormValues } from '@/features/posts/types/post-form-values';
 import { USER_GUIDE_KEYS } from '@/shared/constants/local-storage';
 import { useOverlayContext } from '@/shared/context/overlay-context';
@@ -12,7 +12,7 @@ import type { ValueOf } from '@/shared/type/types';
 import Tab from '@/shared/ui/tab';
 import { SiteHeader } from '@/widgets/header';
 import { useEffect, useState } from 'react';
-import ConsultingPostForm from '@/features/posts/ui/consulting-post-form/consulting-post-form';
+import ConsultingPostForm from '@/features/posts/ui/consulting-form/consulting-post-form/consulting-post-form';
 import useShowModal from '@/shared/ui/hooks/use-show-modal';
 import useShowReloadConsultingPostModal from '@/features/posts/hooks/use-show-reload-consulting-post-modal';
