+++ src/app/posts/page.tsx
@@ -5,7 +5,7 @@ import useGetPosts from '@/features/posts/api/use-get-posts';
 import { POST_TABS, type POST_TAB_VALUES } from '@/features/posts/constants/post-tabs';
 import { getPostListTabs } from '@/features/posts/lib/get-post-list-tabs';
 import type { PostListTab } from '@/features/posts/types/post-list-tab';
-import PostList from '@/features/posts/ui/post-list';
+import PostList from '@/features/posts/ui/post-list/post-list';
 import { WritePostButton } from '@/features/posts/ui/write-post-button';
 import { ROUTES } from '@/shared';
 import { useAuthContext } from '@/features/auth/context/auth-context';
