+++ src/features/posts/ui/consulting-form/consulting-form-option-list.tsx
@@ -1,5 +1,5 @@
 import { useFormContext, useWatch } from 'react-hook-form';
-import type { ConsultingFormOption } from '../types/consulting-form-option';
+import type { ConsultingFormOption } from '../../types/consulting-form-option';
 import ConsultingFormOptionListItem from './consulting-form-option-list-item';
 
 type ConsultingFormOptionListProps = {
