+++ src/features/posts/ui/consulting-form/consulting-post-form/consulting-post-form.tsx
@@ -3,11 +3,11 @@ import type { KeyOf } from '@/shared/type/types';
 import MultiStepForm from '@/shared/ui/multi-step-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { FormProvider, useForm } from 'react-hook-form';
-import { CONSULTING_POST_FORM_FIELD_NAME } from '../../constants/consulting-post-form-field-name';
+import { CONSULTING_POST_FORM_FIELD_NAME } from '../../../constants/consulting-post-form-field-name';
 import {
   consultingPostFormSchema,
   type ConsultingPostFormValues,
-} from '../../types/consulting-post-form-values';
+} from '../../../types/consulting-post-form-values';
 import ConsultingPostFormStep1 from './consulting-post-form-step-1';
 import ConsultingPostFormStep2 from './consulting-post-form-step-2';
 import ConsultingPostFormStep3 from './consulting-post-form-step-3';
