+++ src/features/posts/ui/consulting-form/consulting-response-form/consulting-response-form-price-form.tsx
@@ -3,8 +3,8 @@ import FormItem from '@/shared/ui/form-item';
 import { Input } from '@/shared/ui/input';
 import { useState } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
-import type { ConsultingResponseFormValues } from '../../types/consulting-response-form-values';
-import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../constants/consulting-response-form-field-name';
+import type { ConsultingResponseFormValues } from '../../../types/consulting-response-form-values';
+import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../../constants/consulting-response-form-field-name';
 
 type FormValue = {
   operationName: string;
