+++ src/features/posts/ui/consulting-form/consulting-response-form/consulting-response-form-step-4.tsx
@@ -1,6 +1,6 @@
 import { useFormContext, useWatch } from 'react-hook-form';
-import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../constants/consulting-response-form-field-name';
-import type { ConsultingResponseFormValues } from '../../types/consulting-response-form-values';
+import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../../constants/consulting-response-form-field-name';
+import type { ConsultingResponseFormValues } from '../../../types/consulting-response-form-values';
 import ConsultingFormOptionList from '../consulting-form-option-list';
 import ConsultingResponseFormOptionNeedConsultation from './consulting-response-form-option-need-consultation';
