+++ src/features/posts/ui/consulting-form/consulting-response-form/consulting-response-form-step-5.tsx
@@ -2,16 +2,16 @@ import { Textarea } from '@/shared';
 import FormItem from '@/shared/ui/form-item';
 import ImageUploaderList from '@/shared/ui/image-uploader-list';
 import { useFormContext, useWatch } from 'react-hook-form';
-import { CONSULTING_POST_FORM_FIELD_NAME } from '../../constants/consulting-post-form-field-name';
-import type { ConsultingPostFormValues } from '../../types/consulting-post-form-values';
+import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../../constants/consulting-response-form-field-name';
+import type { ConsultingResponseFormValues } from '../../../types/consulting-response-form-values';
 
 const MAX_IMAGE_COUNT = 6;
 
-export default function ConsultingPostFormStep4() {
-  const { setValue, control, register } = useFormContext<ConsultingPostFormValues>();
+export default function ConsultingResponseFormStep5() {
+  const { register, setValue, control } = useFormContext<ConsultingResponseFormValues>();
 
   const currentOptionValue = useWatch({
-    name: CONSULTING_POST_FORM_FIELD_NAME.option4,
+    name: CONSULTING_RESPONSE_FORM_FIELD_NAME.option5,
     control,
   });
   const currentImages = currentOptionValue?.images ?? [];
@@ -19,7 +19,7 @@ export default function ConsultingPostFormStep4() {
   const handleImageUpload = (file: File) => {
     const newImages = [...currentImages, file];
 
-    setValue(CONSULTING_POST_FORM_FIELD_NAME.option4, {
+    setValue(CONSULTING_RESPONSE_FORM_FIELD_NAME.option5, {
       images: newImages,
       description: currentOptionValue?.description,
     });
@@ -29,7 +29,7 @@ export default function ConsultingPostFormStep4() {
     const newImages = [...currentImages];
     newImages.splice(index, 1);
 
-    setValue(CONSULTING_POST_FORM_FIELD_NAME.option4, {
+    setValue(CONSULTING_RESPONSE_FORM_FIELD_NAME.option5, {
       images: newImages,
       description: currentOptionValue?.description,
     });
@@ -45,10 +45,10 @@ export default function ConsultingPostFormStep4() {
           maxImageCount={MAX_IMAGE_COUNT}
         />
       </FormItem>
-      <FormItem label="추구미 설명">
+      <FormItem label="스타일 설명">
         <Textarea
-          {...register(`${CONSULTING_POST_FORM_FIELD_NAME.option4}.description`)}
-          placeholder="어떤 고민이 있는지 상세히 설명해주세요"
+          {...register(`${CONSULTING_RESPONSE_FORM_FIELD_NAME.option5}.description`)}
+          placeholder="추천하는 스타일을 설명해보세요"
           className="min-h-38"
           hasBorder
         />
