+++ src/features/posts/ui/consulting-form/consulting-response-form/consulting-response-form-step-6.tsx
@@ -1,7 +1,7 @@
 import { useFormContext, useWatch } from 'react-hook-form';
 import ConsultingResponseFormPriceForm from './consulting-response-form-price-form';
-import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../constants/consulting-response-form-field-name';
-import type { ConsultingResponseFormValues } from '../../types/consulting-response-form-values';
+import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../../constants/consulting-response-form-field-name';
+import type { ConsultingResponseFormValues } from '../../../types/consulting-response-form-values';
 import ConsultingFormInputResultListItem from '../consulting-form-input-result-list-item';
 
 export default function ConsultingResponseFormStep6() {
