+++ src/features/posts/ui/consulting-form/consulting-response-form/consulting-response-form-step-7.tsx
@@ -1,8 +1,8 @@
 import { Textarea } from '@/shared';
 import FormItem from '@/shared/ui/form-item';
 import { useFormContext } from 'react-hook-form';
-import type { ConsultingResponseFormValues } from '../../types/consulting-response-form-values';
-import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../constants/consulting-response-form-field-name';
+import type { ConsultingResponseFormValues } from '../../../types/consulting-response-form-values';
+import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../../constants/consulting-response-form-field-name';
 
 export default function ConsultingResponseFormStep7() {
   const { register } = useFormContext<ConsultingResponseFormValues>();
