+++ src/features/posts/ui/consulting-form/consulting-response-form/consulting-response-form.tsx
@@ -2,8 +2,8 @@ import type { FormStep } from '@/shared/type/form-step';
 import type { KeyOf } from '@/shared/type/types';
 import MultiStepForm from '@/shared/ui/multi-step-form';
 import { FormProvider, useForm } from 'react-hook-form';
-import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../constants/consulting-response-form-field-name';
-import type { ConsultingResponseFormValues } from '../../types/consulting-response-form-values';
+import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../../constants/consulting-response-form-field-name';
+import type { ConsultingResponseFormValues } from '../../../types/consulting-response-form-values';
 import ConsultingResponseFormStep1 from './consulting-response-form-step-1';
 import ConsultingResponseFormStep2 from './consulting-response-form-step-2';
 import ConsultingResponseFormStep3 from './consulting-response-form-step-3';
