+++ src/features/posts/ui/post-detail/post-detail-image.tsx
