+++ src/features/posts/ui/post-detail/post-detail-item.tsx
