+++ src/features/posts/ui/post-list/post-list-item.tsx
